[REF] split bank models
