+++ l10n_ch_base_bank/models/__init__.py
@@ -1,2 +1,3 @@
 from . import bank
+from . import partner_bank
 from . import invoice
