+++ l10n_ch_base_bank/postfinance.py
@@ -1,0 +1,5 @@
+# Copyright 2020 Camptocamp
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+CLEARING = "09000"
+BIC = "POFICHBEXXX"
